backend/requirements.txt
Django==3.2.5
django-filter==22.1
djangorestframework==3.14.0
Pillow==8.3.1
djoser
reportlab
gunicorn==20.0.4
psycopg2-binary==2.8.6
drf_base64